# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  request.Login:
    properties:
      email:
        type: string
      password:
        maxLength: 30
        minLength: 5
        type: string
      phone:
        maxLength: 10
        minLength: 10
        type: string
      user_name:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - password
    type: object
  request.OTPLogin:
    properties:
      email:
        type: string
      phone:
        maxLength: 10
        minLength: 10
        type: string
      user_name:
        maxLength: 16
        minLength: 3
        type: string
    type: object
  request.OTPVerify:
    properties:
      otp:
        maxLength: 8
        minLength: 4
        type: string
      otp_id:
        type: string
    required:
    - otp
    type: object
...
  request.RefreshToken:
    properties:
      refresh_token:
        minLength: 10
        type: string
    type: object
...
  request.UserSignUp:
    properties:
      age:
        type: integer
      confirm_password:
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 1
        type: string
      password:
        type: string
      phone:
        maxLength: 10
        minLength: 10
        type: string
      user_name:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - age
    - confirm_password
    - email
    - first_name
    - last_name
    - password
    - phone
    - user_name
    type: object
...
  response.OTPResponse:
    properties:
      otp_id:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact:
    email: okeke98@gmail.com
    name: For API Support
  description: 'Backend API built with Golang using Clean Code architecture. \nGithub:
    [https://github.com/okekefrancis112/ecommerce-gin].'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-commerce Application Backend API
paths:
...
  /admin/auth/renew-access-token:
    post:
      description: API for admin to renew access token using refresh token
      operationId: AdminRenewAccessToken
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RefreshToken'
      responses:
        "200":
          description: Successfully generated access token using refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Refresh token blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: No session found for the given refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "410":
          description: Refresh token expired
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed generate access token
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Renew Access Token (Admin)
      tags:
      - Admin Authentication
  /admin/auth/sign-in:
    post:
      description: API for admin to login with password
      operationId: AdminLogin
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      responses:
        "200":
          description: Successfully logged in
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TokenResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Wrong password
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Admin not exist with this details
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to login
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login with password (Admin)
      tags:
      - Admin Authentication
...
  /auth/renew-access-token:
    post:
      description: API for user to renew access token using refresh token
      operationId: UserRenewAccessToken
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RefreshToken'
      responses:
        "200":
          description: Successfully generated access token using refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Refresh token blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: No session found for the given refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "410":
          description: Refresh token expired
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed generate access token
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Renew Access Token (User)
      tags:
      - User Authentication
  /auth/sign-in:
    post:
      description: API for user to login with email | phone | user_name with password
      operationId: UserLogin
      parameters:
      - description: Login Details
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      responses:
        "200":
          description: Successfully logged in
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TokenResponse'
              type: object
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: User not exist with given login credentials
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked by admin
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to login
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login with password (User)
      tags:
      - User Authentication
  /auth/sign-in/otp/send:
    post:
      description: 'API for user to send otp for login enter email | phone | user_name
        : otp will send to user registered number'
      operationId: UserLoginOtpSend
      parameters:
      - description: Login credentials
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.OTPLogin'
      responses:
        "200":
          description: Successfully otp send to user's registered number
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid Otp
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: User not exist with given login credentials
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked by admin
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to send otp
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login with Otp send (User)
      tags:
      - User Authentication
  /auth/sign-in/otp/verify:
    post:
      description: API for user to verify otp
      operationId: UserLoginOtpVerify
      parameters:
      - description: Otp Verify Details
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.OTPVerify'
      responses:
        "200":
          description: Successfully user logged in
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TokenResponse'
              type: object
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Otp not matched
          schema:
            $ref: '#/definitions/response.Response'
        "410":
          description: Otp Expired
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to verify otp
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login with Otp verify (User)
      tags:
      - User Authentication
  /auth/sign-up:
    post:
      description: API for user to register a new account
      operationId: UserSignUp
      parameters:
      - description: Input Fields
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserSignUp'
      responses:
        "200":
          description: Successfully account created and otp send to registered number
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.OTPResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: A verified user already exist with given user credentials
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to signup
          schema:
            $ref: '#/definitions/response.Response'
      summary: Signup (User)
      tags:
      - User Authentication
  /auth/sign-up/verify:
    post:
      description: API for user to verify otp on sign up
      operationId: UserSignUpVerify
      parameters:
      - description: Otp Verify Details
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.OTPVerify'
      responses:
        "200":
          description: Successfully otp verified for user sign up
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TokenResponse'
              type: object
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Otp not matched
          schema:
            $ref: '#/definitions/response.Response'
        "410":
          description: Otp Expired
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to verify otp
          schema:
            $ref: '#/definitions/response.Response'
      summary: UserSingUp verify OTP  (User)
      tags:
      - User Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'Add prefix of Bearer before  token Ex: "Bearer token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"